+++ android/src/main/java/com/clevertap/clevertap_plugin/CleverTapPlugin.java
@@ -33,6 +33,11 @@ import com.clevertap.android.sdk.product_config.CTProductConfigListener;
 import com.clevertap.android.sdk.pushnotification.CTPushNotificationListener;
 import com.clevertap.android.sdk.pushnotification.PushConstants.PushType;
 import com.clevertap.android.sdk.pushnotification.amp.CTPushAmpListener;
+import com.clevertap.android.sdk.variables.CTVariableUtils;
+import com.clevertap.android.sdk.variables.Var;
+import com.clevertap.android.sdk.variables.callbacks.FetchVariablesCallback;
+import com.clevertap.android.sdk.variables.callbacks.VariableCallback;
+import com.clevertap.android.sdk.variables.callbacks.VariablesChangedCallback;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -87,6 +92,8 @@ public class CleverTapPlugin implements ActivityAware,
 
     private Context context;
 
+    public static Map<String, Object> variables = new HashMap<>();
+
     /**
      * Plugin registration.
      */
@@ -109,7 +116,7 @@ public class CleverTapPlugin implements ActivityAware,
     @Override
     public void onShow(CTInAppNotification ctInAppNotification) {
         invokeMethodOnUiThread("inAppNotificationShow",
-                Utils.jsonObjectToMap(ctInAppNotification.getJsonDescription()));
+                Utils.jsonToMap(ctInAppNotification.getJsonDescription()));
     }
 
     @Override
@@ -187,8 +194,12 @@ public class CleverTapPlugin implements ActivityAware,
     }
 
     @Override
-    public void onInboxItemClicked(final CTInboxMessage message) {
-        invokeMethodOnUiThread("onInboxMessageClick", Utils.jsonObjectToMap(message.getData()));
+    public void onInboxItemClicked(CTInboxMessage message, int contentPageIndex, int buttonIndex) {
+        Map<String, Object> payloadMap = new HashMap<>();
+        payloadMap.put("data", Utils.jsonToMap(message.getData()));
+        payloadMap.put("contentPageIndex", contentPageIndex);
+        payloadMap.put("buttonIndex", buttonIndex);
+        invokeMethodOnUiThread("onInboxMessageClick", payloadMap);
     }
 
     @Override
@@ -200,6 +211,9 @@ public class CleverTapPlugin implements ActivityAware,
     @Override
     public void onMethodCall(MethodCall call, @NonNull Result result) {
         switch (call.method) {
+            case "setLibrary": {
+                setLibrary(call, result);
+            }
             case "setDebugLevel": {
                 int debugLevelValue = call.argument("debugLevel");
                 CleverTapAPI.setDebugLevel(debugLevelValue);
@@ -469,6 +483,10 @@ public class CleverTapPlugin implements ActivityAware,
                 showInbox(call, result);
                 break;
             }
+            case "dismissInbox": {
+                dismissInbox(result);
+                break;
+            }
             case "getInboxMessageCount": {
                 getInboxMessageCount(result);
                 break;
@@ -497,8 +515,18 @@ public class CleverTapPlugin implements ActivityAware,
                 break;
             }
 
+            case "deleteInboxMessagesForIds": {
+                deleteInboxMessagesForIds(call, result);
+                break;
+            }
+
             case "markReadInboxMessageForId": {
                 markReadInboxMessageForId(call, result);
+                break;
+            }
+
+            case "markReadInboxMessagesForIds": {
+                markReadInboxMessagesForIds(call, result);
                 break;
             }
 
@@ -509,6 +537,47 @@ public class CleverTapPlugin implements ActivityAware,
 
             case "pushInboxNotificationViewedEventForId": {
                 pushInboxNotificationViewedEventForId(call, result);
+                break;
+            }
+
+            case "syncVariables": {
+                syncVariables(result);
+                break;
+            }
+
+            //no-op for android, methods only for iOS.
+            case "syncVariablesinProd": {
+                Log.d(TAG, "syncVariablesinProd" + ERROR_IOS);
+                break;
+            }
+
+            case "defineVariables": {
+                defineVariables(call, result);
+                break;
+            }
+
+            case "fetchVariables": {
+                fetchVariables(result);
+                break;
+            }
+
+            case "getVariable": {
+                getVariable(call, result);
+                break;
+            }
+
+            case "getVariables": {
+                getVariables(result);
+                break;
+            }
+
+            case "onVariablesChanged": {
+                onVariablesChanged(result);
+                break;
+            }
+
+            case "onValueChanged": {
+                onValueChanged(call, result);
                 break;
             }
 
@@ -611,6 +680,173 @@ public class CleverTapPlugin implements ActivityAware,
 
     }
 
+    /**************************************************
+     *  Product Experience Remote Config methods starts
+     *************************************************/
+    private void syncVariables(Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            cleverTapAPI.syncVariables();
+            result.success(null);
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    public void defineVariables(MethodCall call, Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            Map<String, Object> variablesMap = call.argument("variables");
+            for (Map.Entry<String, Object> entry : variablesMap.entrySet()) {
+                String key = entry.getKey();
+                Object value = entry.getValue();
+                variables.put(key, cleverTapAPI.defineVariable(key, value));
+            }
+            result.success(null);
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    public void fetchVariables(Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            cleverTapAPI.fetchVariables(new FetchVariablesCallback() {
+                @Override
+                public void onVariablesFetched(final boolean isSuccess) {
+                    result.success(isSuccess);
+                }
+            });
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    public void getVariable(MethodCall call, Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            try {
+                String key = call.argument("name");
+                result.success(getVariableValue(key));
+            } catch (Exception e) {
+                result.error(TAG, "Unable to get the variable value: " + e.getLocalizedMessage(), null);
+            }
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    public void getVariables(Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            try {
+                result.success(getVariablesValues());
+            } catch (Exception e) {
+                result.error(TAG, "Unable to get the variable value: " + e.getLocalizedMessage(), null);
+            }
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    public void onVariablesChanged(Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            cleverTapAPI.addVariablesChangedCallback(new VariablesChangedCallback() {
+                @Override
+                public void variablesChanged() {
+                    invokeMethodOnUiThread("onVariablesChanged", getVariablesValues());
+                }
+            });
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    public void onValueChanged(MethodCall call, Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            String name = call.argument("name");
+            if (variables.containsKey(name)) {
+
+                Var<Object> var = (Var<Object>) variables.get(name);
+                if (var != null) {
+                    var.addValueChangedCallback(new VariableCallback<Object>() {
+                        @SuppressLint("RestrictedApi")
+                        @Override
+                        public void onValueChanged(final Var<Object> variable) {
+                            Map<String, Object> variablesMap = new HashMap<>();
+                            try {
+                                variablesMap = getVariableValueAsWritableMap(name);
+                                result.success(variablesMap);
+                            } catch (Exception e) {
+                                result.error(TAG, "Unable to handle onValueChanged callback: " + e.getLocalizedMessage(), null);
+                            }
+                            invokeMethodOnUiThread("onValueChanged", variablesMap);
+                        }
+                    });
+                } else {
+                    Log.d(TAG, "Variable value with name = " + name + " contains null value. Not setting onValueChanged callback.");
+                }
+            } else {
+                Log.e(TAG, "Variable name = " + name + " does not exist. Make sure you set variable first.");
+            }
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    /************************************************
+     *  Product Experience Remote Config methods ends
+     ************************************************/
+
+    @SuppressLint("RestrictedApi")
+    private Object getVariableValue(String name) {
+        if (variables.containsKey(name)) {
+            Var<?> variable = (Var<?>) variables.get(name);
+            Object variableValue = variable.value();
+            Object value;
+            switch (variable.kind()) {
+                case CTVariableUtils.DICTIONARY:
+                    //value = CleverTapUtils.MapUtil.toWritableMap((Map<String, Object>) variableValue);
+                    return variableValue;
+                default:
+                    value = variableValue;
+            }
+            return value;
+        }
+        throw new IllegalArgumentException(
+                "Variable name = " + name + " does not exist. Make sure you set variable first.");
+    }
+
+    private Map<String, Object> getVariablesValues() {
+        Map<String, Object> variablesMapObject = new HashMap<>();
+        for (Map.Entry<String, Object> entry : variables.entrySet()) {
+            String key = entry.getKey();
+            Var<?> variable = (Var<?>) entry.getValue();
+
+            Map<String, Object> variableWritableMap = CleverTapTypeUtils.MapUtil.addValue(key, variable.value());
+            variablesMapObject.putAll(variableWritableMap);
+        }
+        return variablesMapObject;
+    }
+
+    @SuppressLint("RestrictedApi")
+    private Map<String, Object> getVariableValueAsWritableMap(String name) {
+        if (variables.containsKey(name)) {
+            Var<?> variable = (Var<?>) variables.get(name);
+            Object variableValue = variable.value();
+            return CleverTapTypeUtils.MapUtil.addValue(name, variable.value());
+        }
+        throw new IllegalArgumentException(
+                "Variable name = " + name + " does not exist.");
+    }
+
+    @SuppressLint("RestrictedApi")
+    private void setLibrary(MethodCall call, Result result) {
+        String libName = call.argument("libName");
+        int libVersion = call.argument("libVersion");
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            cleverTapAPI.setCustomSdkVersion(libName, libVersion);
+            result.success(null);
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
     @Override
     public void onNotificationClickedPayloadReceived(HashMap<String, Object> hashMap) {
         invokeMethodOnUiThread("pushClickedPayloadReceived", hashMap);
@@ -628,7 +864,7 @@ public class CleverTapPlugin implements ActivityAware,
 
     @Override
     public void profileDataUpdated(JSONObject updates) {
-        invokeMethodOnUiThread("profileDataUpdated", Utils.jsonObjectToMap(updates));
+        invokeMethodOnUiThread("profileDataUpdated", Utils.jsonToMap(updates));
     }
 
     @Override
@@ -787,6 +1023,40 @@ public class CleverTapPlugin implements ActivityAware,
         }
     }
 
+    public void deleteInboxMessagesForIds(MethodCall call, Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            ArrayList<String> messageIds = call.argument("messageIds");
+            cleverTapAPI.deleteInboxMessagesForIDs(messageIds);
+            result.success(null);
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    private void markReadInboxMessageForId(MethodCall call, Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            String messageId = call.argument("messageId");
+            if (messageId == null || messageId.isEmpty()) {
+                result.error(TAG, ERROR_MSG_ID, null);
+                return;
+            }
+            cleverTapAPI.markReadInboxMessage(messageId);
+            result.success(null);
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    public void markReadInboxMessagesForIds(MethodCall call, Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            ArrayList<String> messageIds = call.argument("messageIds");
+            cleverTapAPI.markReadInboxMessagesForIDs(messageIds);
+            result.success(null);
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
     private void enableDeviceNetworkInfoReporting(MethodCall call, Result result) {
         boolean value = call.argument("value");
         if (isCleverTapNotNull(cleverTapAPI)) {
@@ -893,7 +1163,7 @@ public class CleverTapPlugin implements ActivityAware,
             if (cleverTapAPI.getDisplayUnitForId(unitId) != null) {
                 JSONObject displayUnit = cleverTapAPI.getDisplayUnitForId(unitId).getJsonObject();
                 if (displayUnit != null) {
-                    result.success(Utils.jsonObjectToMap(displayUnit));
+                    result.success(Utils.jsonToMap(displayUnit));
                 }
             } else {
                 result.error(TAG, "Display Unit is NULL", null);
@@ -947,7 +1217,7 @@ public class CleverTapPlugin implements ActivityAware,
             }
             CTInboxMessage inboxMessage = cleverTapAPI.getInboxMessageForId(messageId);
             if (inboxMessage != null) {
-                result.success(Utils.jsonObjectToMap(inboxMessage.getData()));
+                result.success(Utils.jsonToMap(inboxMessage.getData()));
             }
         } else {
             result.error(TAG, ERROR_MSG, null);
@@ -1082,20 +1352,6 @@ public class CleverTapPlugin implements ActivityAware,
         return cleverTapAPI != null;
     }
 
-    private void markReadInboxMessageForId(MethodCall call, Result result) {
-        if (isCleverTapNotNull(cleverTapAPI)) {
-            String messageId = call.argument("messageId");
-            if (messageId == null || messageId.isEmpty()) {
-                result.error(TAG, ERROR_MSG_ID, null);
-                return;
-            }
-            cleverTapAPI.markReadInboxMessage(messageId);
-            result.success(null);
-        } else {
-            result.error(TAG, ERROR_MSG, null);
-        }
-    }
-
     private void onUserLogin(MethodCall call, Result result) {
         Map<String, Object> profile = Utils.dartMapToProfileMap(call.argument("profile"));
         if (isCleverTapNotNull(cleverTapAPI)) {
@@ -1107,10 +1363,10 @@ public class CleverTapPlugin implements ActivityAware,
     }
 
     private void processPushNotification(MethodCall call, Result result) {
-        JSONObject extras = call.argument("extras");
+        Map<String, Object> extras = call.argument("extras");
         if (isCleverTapNotNull(cleverTapAPI)) {
             try {
-                CleverTapAPI.processPushNotification(context, Utils.jsonToBundle(extras));
+                CleverTapAPI.processPushNotification(context, Utils.jsonToBundle(new JSONObject(extras)));
             } catch (JSONException e) {
                 result.error(TAG, "Unable to render notification due to JSONException - " + e.getLocalizedMessage(),
                         null);
@@ -1537,7 +1793,7 @@ public class CleverTapPlugin implements ActivityAware,
         this.dartToNativeMethodChannel = getMethodChannel("clevertap_plugin/dart_to_native", messenger, registrar);
         if (nativeToDartMethodChannel == null) {
             // set nativeToDartMethodChannel channel once and it has to be static field
-            // and per https://github.com/firebase/flutterfire/issues/9689 because multiple
+            // as per https://github.com/firebase/flutterfire/issues/9689 because multiple
             // instances of the CleverTap plugin can be created in case onBackgroundMessage handler
             // of FCM plugin.
             nativeToDartMethodChannel = getMethodChannel("clevertap_plugin/native_to_dart", messenger, registrar);
@@ -1557,7 +1813,6 @@ public class CleverTapPlugin implements ActivityAware,
             this.cleverTapAPI.setCTFeatureFlagsListener(this);
             this.cleverTapAPI.setCTProductConfigListener(this);
             this.cleverTapAPI.setCTPushAmpListener(this);
-            this.cleverTapAPI.setLibrary("Flutter");
             this.cleverTapAPI.registerPushPermissionNotificationResponseListener(this);
         }
     }
@@ -1569,6 +1824,15 @@ public class CleverTapPlugin implements ActivityAware,
         styleConfig = Utils.jsonToStyleConfig(styleConfigJson);
         if (isCleverTapNotNull(cleverTapAPI)) {
             cleverTapAPI.showAppInbox(styleConfig);
+            result.success(null);
+        } else {
+            result.error(TAG, ERROR_MSG, null);
+        }
+    }
+
+    private void dismissInbox(Result result) {
+        if (isCleverTapNotNull(cleverTapAPI)) {
+            cleverTapAPI.dismissAppInbox();
             result.success(null);
         } else {
             result.error(TAG, ERROR_MSG, null);